handled the case where None is explicitly returned from a function

+++ test/samples/compiler/function_call6.py
@@ -1,4 +1,4 @@
-def my_function(x: int = 1, y: float = 2.0) -> int:
+def my_function(x: int = 1, y: float = 2.0) -> float:
     return x / y
